refactor(metro): centralize and reuse metro configuration logic
