+++ e2e-app/package.json
@@ -48,6 +48,7 @@
     "eas-cli": "^16.18.0",
     "jest": "^29.2.1",
     "jest-expo": "~53.0.10",
+    "metro-config": "^0.82.5",
     "react-test-renderer": "18.3.1",
     "semver": "^7.8.1",
     "serve": "^14.2.4",
